# backend/well-known/openapi.yaml
openapi: 3.1.0
info:
  title: Student Manager API
  version: 1.0.0
  description: API for reading and modifying students stored in a JSON file.

servers:
  - url: https://YOUR_DEPLOYED_BACKEND_URL/api

paths:
  /students:
    get:
      summary: Get all students
      operationId: getStudents
      responses:
        "200":
          description: List of all students

    post:
      summary: Add a new student
      operationId: addStudent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                roll:
                  type: integer
                name:
                  type: string
                grade:
                  type: string
              required:
                - roll
                - name
                - grade
      responses:
        "200":
          description: Student added

  /students/{roll}:
    get:
      summary: Get a student by roll
      operationId: getStudentByRoll
      parameters:
        - name: roll
          in: path
          schema:
            type: integer
          required: true
      responses:
        "200":
          description: Student found
        "404":
          description: Student not found

    put:
      summary: Update student by roll
      operationId: updateStudent
      parameters:
        - name: roll
          in: path
          schema:
            type: integer
          required: true
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                grade:
                  type: string
      responses:
        "200":
          description: Student updated
        "404":
          description: Student not found

    delete:
      summary: Delete student by roll
      operationId: deleteStudent
      parameters:
        - name: roll
          in: path
          schema:
            type: integer
          required: true
      responses:
        "200":
          description: Student deleted
        "404":
          description: Student not found